Fix all CI code quality issues

+++ tasks/sosnina_a_diff_count/mpi/src/ops_mpi.cpp
@@ -43,7 +43,7 @@ bool SosninaADiffCountMPI::RunImpl() {
   std::size_t remainder = total_len % size;
 
   std::size_t start = (rank * block_size) + std::min(static_cast<std::size_t>(rank), remainder);
-  std::size_t end = start + block_size + (static_cast<std::size_t>(rank) < remainder ? 1 : 0);
+  std::size_t end = start + block_size + (rank < static_cast<int>(remainder) ? 1 : 0);
   end = std::min(end, total_len);
 
   int local_diff_count = 0;
